Split SQL Driver specific code into different adapters
